implementing postman protections

+++ app/controllers/friendships_controller.rb
@@ -26,29 +26,36 @@ class FriendshipsController < ApplicationController
     friend = User.find_by(id: params[:friend_id])
     friendee_friendship = user.friendee_friendships.where(friender_id: friend.id)
     friender_friendship = user.friender_friendships.where(friendee_id: friend.id)
-    if friendee_friendship.size > 0
-      friendee_invitations = friendee_friendship.first().get_invitations
-      friendee_attendances = friendee_friendship.first().get_attendances
-      friendee_friendship.each {|el| el.destroy}
-      if friendee_invitations.size > 0
-        friendee_invitations.first.destroy_all
+    if params[:user_id] == session[:user_id]
+      if friendee_friendship.size > 0
+        friendee_invitations = friendee_friendship.first().get_invitations
+        friendee_attendances = friendee_friendship.first().get_attendances
+        friendee_friendship.each {|el| el.destroy}
+        if friendee_invitations.size > 0
+          friendee_invitations.first.destroy_all
+        end
+        if friendee_attendances.size > 0
+          friendee_attendances.first.destroy_all
+        end
+      elsif friender_friendship.size > 0
+        friender_invitations = friender_friendship.first().get_invitations
+        friender_attendances = friender_friendship.first().get_attendances
+        friender_friendship.each {|el| el.destroy}
+        if friender_invitations.size > 0
+          friender_invitations.first.destroy_all
+        end
+        if friender_attendances.size > 0
+          friender_attendances.first.destroy_all
+        end
       end
-      if friendee_attendances.size > 0
-        friendee_attendances.first.destroy_all
-      end
-    elsif friender_friendship.size > 0
-      friender_invitations = friender_friendship.first().get_invitations
-      friender_attendances = friender_friendship.first().get_attendances
-      friender_friendship.each {|el| el.destroy}
-      if friender_invitations.size > 0
-        friender_invitations.first.destroy_all
-      end
-      if friender_attendances.size > 0
-        friender_attendances.first.destroy_all
-      end
+      friends = user.friends
+      attendances = user.attendances
+      receiver_invitations = user.receiver_invitations
+      sender_invitations = user.sender_invitaitons
+      render json: friends
+    else
+      return render json: { error: "Not authorized" }, status: :unauthorized
     end
-    friends = user.friends
-    render json: friends
   end
 
   private
